requirements.txt
uvicorn~=0.11.8
fastapi~=0.61.1
pandas~=1.1.2
matplotlib~=3.3.2
scikit-learn~=0.23.2
scipy~=1.5.2
numpy~=1.19.2
kneed~=0.7.0
imblearn~=0.0
pymongo~=3.11.0
sklearn_pandas~=2.0.2